fix(codegen/plc4go): fixed issue with unnecessary casting when using bitwise operators

+++ plc4go/protocols/cbus/readwrite/model/MeasurementDataChannelMeasurementData.go
@@ -124,7 +124,7 @@ func (m *_MeasurementDataChannelMeasurementData) GetLsb() uint8 {
 ///////////////////////
 
 func (m *_MeasurementDataChannelMeasurementData) GetRawValue() uint16 {
-	return uint16(uint16(uint16(m.GetMsb())<<uint16(uint16(8))) | uint16(m.GetLsb()))
+	return uint16(m.GetMsb()<<uint16(8) | m.GetLsb())
 }
 
 func (m *_MeasurementDataChannelMeasurementData) GetValue() float64 {
@@ -260,7 +260,7 @@ func MeasurementDataChannelMeasurementDataParse(readBuffer utils.ReadBuffer) (Me
 	lsb := _lsb
 
 	// Virtual field
-	_rawValue := uint16(uint16(msb)<<uint16(uint16(8))) | uint16(lsb)
+	_rawValue := msb<<uint16(8) | lsb
 	rawValue := uint16(_rawValue)
 	_ = rawValue
